ExemplosMod29Projeto3/src.java.main/br/com/calleb/domain/Produto.java
package br.com.calleb.domain;

/**
 * Description of Produto
 * Created by calle on 01/09/2023.
 */
public class Produto {

    private Long id;
    private String codigo;
    private String nome;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getCodigo() {
        return codigo;
    }

    public void setCodigo(String codigo) {
        this.codigo = codigo;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }
}
